Add new images and fix image file extension
This commit adds several new images to the journal entries and fixes an incorrect file extension for one of the images. The changes include adding images of a beholder zombie, Ilvarra, a purple worm, ettins, kuo-toa, and a darklake raft battle. Additionally, the commit corrects the file extension for the beholder-zombie image from .jpg to .png.