Add spell-check words, fix typos and add details to journal entries
This commit adds a new file with spell-check words. It also fixes some typos in the first and twentieth journal entries. In the first entry, it changes "wagonload" to "wagon-load". In the twentieth entry, it capitalizes "Uglies" and adds "polearm" after "Mancatcher". Additionally, this commit adds more details to the first and twentieth journal entries by specifying that Gundren Rockseeker is a dwarf and Sildar Hallwinter is a warrior escort.