feat: Add journal entry for Thirty-Fourth Session
This commit adds a new journal entry for the Thirty-Fourth Session. The session begins with the party boarding an airship to Gauntlgrym and meeting various characters along the way. They discuss their adventures and receive information about the dangers of the underdark. The party also encounters trolls in a battle, where they use their abilities to deal damage. Additionally, there is a minor change in the Thirty-Third Session journal entry, correcting a typo regarding the fate of the Dwarves during an encounter with tentacles.